requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
python-dotenv
pypdf
torch
pysqlite3-binary
langchain-classic